fix: eliminate data race on application mode and reason fields
Fixes #246

Changed mode from gen.ApplicationMode to int32 with atomic operations.
Changed reason from error to atomic.Value for thread-safe access.

This fixes the race between stop() and terminate() during shutdown.
Also fixed race on state field in registerAppRoute() by using atomic load.

+++ node/node.go
@@ -1591,7 +1591,7 @@ func (n *node) ApplicationLoad(app gen.ApplicationBehavior, args ...any) (name g
 		node:     n,
 		behavior: app,
 		state:    int32(gen.ApplicationStateLoaded),
-		mode:     spec.Mode,
+		mode:     int32(spec.Mode),
 	}
 	if _, exist := n.applications.LoadOrStore(spec.Name, a); exist {
 		return spec.Name, gen.ErrTaken
